MessageSchema

+++ fnord-cstable/BinaryFormat.h
@@ -52,7 +52,7 @@ public:
 };
 
 
-enum class ColumnType : uint32_t {
+enum class ColumnType : uint8_t {
   BOOLEAN = 1,
   UINT32_BITPACKED = 10,
   UINT32_PLAIN = 11,
